+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Added a new class on the Python framework: `InputValidation`.
 - New request: Delete a list of groups - `DELETE /agents/groups - Params: {"ids":["id_1",...,"id_n"]}`
 - Input Validation at framework level of `group_id` parameter.
+- New request: Get information about the actual manager node in the cluster - `GET/cluster/node`
 
 ### Fixed
 - When adding a new agent, now it checks that its name is different than manager's name.