+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Improve output of delete and restart agents requests: each ID includes error information.
 - Add requests to manage remote agent upgrades.
 - New request: Get information about the actual manager node in the cluster - `GET/cluster/node`
+- New request: Get information about  all nodes in the cluster - `GET/cluster/nodes`
 - New request: Get information about the status of the synchronized files in the cluster - `GET/cluster/files`
 - New request: Get information about the agents in the cluster - `GET/cluster/agents`
 - Add a selector for the API call to retrieve information about an agent using its ID.